requirements.txt
beautifulsoup4==4.12.3
et-xmlfile==1.1.0
greenlet==3.0.3
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
playwright==1.42.0
pyee==11.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
soupsieve==2.5
typing_extensions==4.10.0
tzdata==2024.1
